removing helper generateAndVerifyClient
