--- packages/helpers/turbo.json
@@ -1,9 +1,0 @@
-{
-  "extends": ["//"],
-  "$schema": "https://turbo.build/schema.v1.json",
-  "pipeline": {
-    "build": {
-      "dependsOn": ["^build", "@asyncapi/generator-react-sdk#build"]
-    }
-  }
-}
