+++ packages/templates/clients/kafka/java/quarkus/components/ClientConsumer.js
@@ -2,7 +2,9 @@ import ConsumeEvent from './ConsumeEvent';
 import { ConsumerFields } from './ConsumerFields';
 import { Text } from '@asyncapi/generator-react-sdk';
 
-export default function ClientConsumer({ className }) {
+export default function ClientConsumer({ className, eventName }) {
+  const formattedEventName = eventName ? eventName.replace(/^\w/, (c) => c.toUpperCase()) : eventName;
+
   return (
     <Text newLines={2}>
       <Text newLines={2}>
@@ -11,7 +13,7 @@ export default function ClientConsumer({ className }) {
 public class ${className}{`}
       </Text>
       <ConsumerFields clientName={className}/>
-      <ConsumeEvent/>
+      <ConsumeEvent eventName={formattedEventName}/>
       
     </Text>
   );
